requirements.txt
aiogram~=2.17.1
aiogram_dialog ~=1.3.0
python-dotenv~=0.19.2
requests